yolo-code-assistant/pyproject.toml
[project]
name = "yolo-code-assistant"
version = "0.1.0"
description = "A RAG-based code assistant for Ultralytics YOLO codebase"
readme = "README.md"
authors = [
    { name = "gauravlochab", email = "gauravlochab1997@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "gradio>=4.0.0",
    "pymongo[srv]>=4.6.0",
    "sentence-transformers>=2.2.0",
    "openai>=1.0.0",
    "tree-sitter>=0.20.0",
    "tree-sitter-python>=0.20.0",
    "GitPython>=3.1.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0", 
    "ruff>=0.1.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[project.scripts]
yolo-code-assistant = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
